Store userdata in 64-bit compatible way

+++ Code/Commando/dlgloadspgame.cpp
@@ -246,9 +246,9 @@ LoadSPGameMenuClass::Build_List (const char *search_string, int start_index)
 				file_path += "\\";
 				file_path += find_info.cFileName;
 				
-				list_ctrl->Set_Entry_Data (item_index, 0, (uint32)new FILETIME(local_time));
-				list_ctrl->Set_Entry_Data (item_index, 1, (uint32)new StringClass(file_path));
-				list_ctrl->Set_Entry_Data (item_index, 2, (uint32)new StringClass(find_info.cFileName));
+				list_ctrl->Set_Entry_Data (item_index, 0, (uintptr_t)new FILETIME(local_time));
+				list_ctrl->Set_Entry_Data (item_index, 1, (uintptr_t)new StringClass(file_path));
+				list_ctrl->Set_Entry_Data (item_index, 2, (uintptr_t)new StringClass(find_info.cFileName));
 			}
 		}
 	}
