+++ Code/Commando/dlgmplangamelist.cpp
@@ -418,7 +418,7 @@ MPLanGameListMenuClass::Update_Game_List (void)
 			//
 			//	Associate the channel data inside with the entry
 			//
-			list_ctrl->Set_Entry_Data (item_index, 0, (uint32)channel);
+			list_ctrl->Set_Entry_Data (item_index, 0, (uintptr_t)channel);
 			channel->Add_Ref ();
 
 			//
